Mire.jl Documentation

Mire.jl Documentation

TODO Introduction

Low level functions

Functions for low level control of the problem.

Mire.mat_force — Function.

mat_force(N::Integer, vs, forcefun::Function, a::T, b::T, c::T, args...) where T

Allocates new matrix A and fills elements by calling matforcegalerkin!(A,vs,N,forcefun,a,b,c, args...).

Cached version: mat_force(N,cmat,vs,forcefun,a,b,c, args...)

where cmat[i,j,k] contains the integrals of monomials xⁱyʲzᵏ.

#Arguments:

N: DESCRIPTION

vs: DESCRIPTION

forcefun: DESCRIPTION

a: DESCRIPTION

b: DESCRIPTION

c: DESCRIPTION

args: DESCRIPTION

source

Mire.mat_force_galerkin! — Function.

mat_force_galerkin!(A::AbstractArray{T, 2}, cmat::Array{T, 3}, vs::Array{Array{P, 1}, 1}, N::Integer, forcefun::Function, a::T, b::T, c::T, args...; kwargs...) where {T, P <: Polynomial{T}}

DOCSTRING

#Arguments:

A: DESCRIPTION

cmat: DESCRIPTION

vs: DESCRIPTION

N: DESCRIPTION

forcefun: DESCRIPTION

a: DESCRIPTION

b: DESCRIPTION

c: DESCRIPTION

args: DESCRIPTION

kwargs: DESCRIPTION

source

Mire.vel — Function.

vel(N,a,b,c)

Compute all velocity basis vectors for a given maximal degree N (Lebovitz 1989, eq. 41-42)

source

Mire.eigenvel — Function.

eigenvel(N,vs,αs,a,b,c; norm=true)

Reconstructs velocity u, following Vidal & Cebron 2017 eq. (3.5).

If norm keyword is set true the velocity is normalised to satisfy $\int u\cdot u dV=1$.

source

Mire.int_ellipsoid_surface — Function.

Integral over the surface of an ellipsoid.

source

Mire.int_monomial_ellipsoid — Function.

int_monomial_ellipsoid(i::BigInt, j::BigInt, k::BigInt, a::T, b::T, c::T; dtype::DataType=BigFloat) where T

DOCSTRING

#Arguments:

i: DESCRIPTION

j: DESCRIPTION

k: DESCRIPTION

a: DESCRIPTION

b: DESCRIPTION

c: DESCRIPTION

dtype: DESCRIPTION

source

Mire.int_polynomial_ellipsoid — Function.

int_polynomial_ellipsoid(p, cmat)

DOCSTRING

source

Mire.inner_product — Function.

inner_product(u,v,a,b,c)

Defines inner product in an ellipsoidal volume $\int\langle u,v\rangle dV$.

source

inner_product(cmat, u, v; thresh=eps())

DOCSTRING

#Arguments:

cmat: DESCRIPTION

u: DESCRIPTION

v: DESCRIPTION

thresh: DESCRIPTION

source

Setting up the eigen problem

TODO: Description how to set up the problem here.

Mire.assemblehd — Function.

assemblehd(N::Int, cmat, a::T, b::T, c::T, Ω) where T

DOCSTRING

#Arguments:

N: DESCRIPTION

cmat: DESCRIPTION

a: DESCRIPTION

b: DESCRIPTION

c: DESCRIPTION

Ω: DESCRIPTION

source

Mire.assemblemhd — Function.

assemblemhd(N::Int, cmat::Array{T, 3}, a::T, b::T, c::T, Ω, b0; kwargs...) where T

DOCSTRING

#Arguments:

N: DESCRIPTION

cmat: DESCRIPTION

a: DESCRIPTION

b: DESCRIPTION

c: DESCRIPTION

Ω: DESCRIPTION

b0: DESCRIPTION

kwargs: DESCRIPTION

source