Functions

Defining the Volume

Mire.Ellipsoid — Type

Ellipsoid{T<:Number} <: Volume{T}

Volume type Ellipsoid. Create by calling Ellipsoid(a,b,c), with a,b,c the semi-axes. a,b,c can be of any Number type.

Examples: Ellipsoid(1.1,1.0,0.9) is a Ellipsoid{Float64}, Ellipsoid(1//1,1//2,1//5) is a Ellispoid{Rational{Int64}}.

Mire.Sphere — Type

Sphere{T<:Number} <: Volume{T}

Volume type Sphere. Create by calling Sphere{T}(), with T any Number type. Default: Sphere(T) gives a Sphere{Float64}(). For other types use, e.g. Sphere{Rational{BigInt}}().

Vector bases

Mire.LebovitzBasis — Type

LebovitzBasis{T<:Number,Vol<:Volume{T}} <: VectorBasis{T,Vol}

Basis of 3-D vector field, so that $\mathbf{u}\cdot\vec{n} = 0$ at $\partial\mathcal{V}$ and $\nabla\cdot\mathbf{u} = 0$ after Lebovitz (1989).

Setting up the problem

Mire.HDProblem — Type

HDProblem{T<:Number,Vol<:Volume{T}} <: MireProblem{T,Vol}

Defines hydrodynamic problem.

Example:

N = 5
Ω = [0,0,1.0]
V = Ellipsoid(1.1,1.0,0.9)
problem = HDProblem(N,V,Ω,LebovitzBasis)

Mire.MHDProblem — Type

MHDProblem{T<:Number,Vol<:Volume{T}} <: MireProblem{T,Vol}

Defines magnetohydrodynamic problem.

Example for a hybrid QG model with 3-D magnetic field with conducting boundary condition and QG velocities:

N = 5
Ω = [0,0,1.0]
a,b,c = 1.1,1.0,0.9
V = Ellipsoid(a,b,c)
B0 = [-y/b^2,x/a^2,0] #Malkus field
problem = MHDProblem(N,V,Ω,B0,QGBasis,ConductingMFBasis)

Mire.assemble! — Function

assemble!(P::HDProblem{T,V}) where {T,V}

Assembles the matrices P.LHS and P.RHS, i.e. projecting the velocity basis P.vbasis on the inertial acceleration and Coriolis force.

assemble!(P::MHDProblem{T,V}) where {T,V}

Assembles the matrices P.LHS and P.RHS, i.e. projecting the velocity and magnetic field bases on the inertial acceleration, Coriolis force, Lorentz force and mgnetic advection.

Low level functions

Functions for low level control of the problem.

Mire.projectforce — Function

projectforce(N::Integer,cmat::Array{T,3},vs_i::Array{Array{P,1},1},vs_j::Array{Array{P,1},1},forcefun::Function,a::T,b::T,c::T, args...) where {T, P <: Polynomial{T}}

Allocates new matrix A and fills elements by calling projectforce!(A,cmat,vsi,vsj,forcefun, args...)

where cmat[i,j,k] contains the integrals of monomials xⁱyʲzᵏ.

#Arguments:

N: maximum monomial degree

vs_i: basis vectors to project on

vs_j: basis vectors used for forcefun

forcefun: function of the force, e.g. coriolis

args: other arguments needed for forcefun

Mire.projectforce! — Function

projectforce!(A::AbstractArray{T, 2}, cmat::Array{T, 3}, vs::Array{Array{P, 1}, 1}, N::Integer, forcefun::Function, a::T, b::T, c::T, args...) where {T, P <: Polynomial{T}}

DOCSTRING

#Arguments:

A: pre-allocated array

cmat: pre-cached monomial integration values

vs: basis vectors

N: maximum monomial degree

forcefun: function of the force, e.g. coriolis

args: other arguments needed for forcefun

Mire.projectforcet! — Function

projectforcet(A::AbstractArray{T,2}, cmat::Array{T,3}, vs_i::Array{Array{P,1},1}, vs_j::Array{Array{P,1},1}, forcefun::Function, args...; n_cache = 10^6) where {T, P <: Polynomial{T}}

Multithreaded projection using pre-cached polynomial terms array. n_cache needs to be sufficiently large for a given polynomial degree of the basis vectors (trial and error for now!).

Mire.inner_product — Function

inner_product(u,v,a,b,c)

Defines inner product in an ellipsoidal volume $\int\langle u,v\rangle dV$.

inner_product(cmat, u, v; thresh=eps())

DOCSTRING

#Arguments:

cmat: DESCRIPTION

u: DESCRIPTION

v: DESCRIPTION

thresh: DESCRIPTION

Mire.uqg — Function

uqg(n::Integer, m::Integer, V::Volume{T}) where T

Generate basis vector $\mathbf{u}=\nabla(h^3x^ny^m)\times\nabla(z/h)$