Functions

Defining the Volume

Mire.Ellipsoid — Type

Ellipsoid{T<:Number} <: Volume{T}

Volume type Ellipsoid. Create by calling Ellipsoid(a,b,c), with a,b,c the semi-axes. a,b,c can be of any Number type.

Examples: Ellipsoid(1.1,1.0,0.9) is a Ellipsoid{Float64}, Ellipsoid(1//1,1//2,1//5) is a Ellispoid{Rational{Int64}}.

Mire.Sphere — Type

Sphere{T<:Number} <: Volume{T}

Volume type Sphere. Create by calling Sphere{T}(), with T any Number type. Default: Sphere(T) gives a Sphere{Float64}(). For other types use, e.g. Sphere{Rational{BigInt}}().

Vector bases

Mire.LebovitzBasis — Type

LebovitzBasis{T<:Number,Vol<:Volume{T}} <: VectorBasis{T,Vol}

Basis of 3-D vector field, so that $\mathbf{u}\cdot\vec{n} = 0$ at $\partial\mathcal{V}$ and $\nabla\cdot\mathbf{u} = 0$ after Lebovitz (1989).

Setting up the problem

Mire.HDProblem — Type

HDProblem{T<:Number,Vol<:Volume{T}}

Defines hydrodynamic problem.

Example:

N = 5
Ω = [0,0,1.0]
V = Ellipsoid(1.1,1.0,0.9)
problem = HDProblem(N,V,Ω,LebovitzBasis)

Mire.MHDProblem — Type

MHDProblem{T<:Number,Vol<:Volume{T}} <: MireProblem{T,Vol}

Defines magnetohydrodynamic problem.

Example for a hybrid QG model with 3-D magnetic field with conducting boundary condition and QG velocities:

N = 5
Ω = [0,0,1.0]
a,b,c = 1.1,1.0,0.9
V = Ellipsoid(a,b,c)
B0 = [-y/b^2,x/a^2,0] #Malkus field
problem = MHDProblem(N,V,Ω,B0,QGBasis,ConductingMFBasis)

Mire.assemble! — Function

assemble!(P::HDProblem{T,V}; threads=false, verbose=false, kwargs...) where {T,V}

Assembles the matrices P.LHS and P.RHS, i.e. projecting the velocity basis P.vbasis on the inertial acceleration and Coriolis force.

assemble!(P::MHDProblem{T,V}) where {T,V}

Assembles the matrices P.LHS and P.RHS, i.e. projecting the velocity and magnetic field bases on the inertial acceleration, Coriolis force, Lorentz force and mgnetic advection.

Low level functions

Functions for low level control of the problem.

Missing docstring.

Missing docstring for projectforce. Check Documenter's build log for details.

Missing docstring.

Missing docstring for projectforce!. Check Documenter's build log for details.

Mire.projectforcet! — Function

projectforcet!(args...)

Multithreaded projection

Mire.inner_product — Function

inner_product(u,v,a,b,c)

Defines inner product in an ellipsoidal volume $\int\langle u,v\rangle dV$.

Mire.uqg — Function

uqg(n::Integer, m::Integer, V::Volume{T}) where T

Generate basis vector $\mathbf{u}=\nabla(h^3x^ny^m)\times\nabla(z/h)$