Functions

Setting up the eigen problem

TODO: Description how to set up the problem here.

Missing docstring.

Missing docstring for assemblehd. Check Documenter's build log for details.

Missing docstring.

Missing docstring for assemblemhd. Check Documenter's build log for details.

Missing docstring.

Missing docstring for assemblemhd_hybrid. Check Documenter's build log for details.

Low level functions

Functions for low level control of the problem.

Mire.projectforce — Function

projectforce(N::Integer,cmat::Array{T,3},vs_i::Array{Array{P,1},1},vs_j::Array{Array{P,1},1},forcefun::Function,a::T,b::T,c::T, args...) where {T, P <: Polynomial{T}}

Allocates new matrix A and fills elements by calling projectforce!(A,cmat,vsi,vsj,forcefun, args...)

where cmat[i,j,k] contains the integrals of monomials xⁱyʲzᵏ.

#Arguments:

N: maximum monomial degree

vs_i: basis vectors to project on

vs_j: basis vectors used for forcefun

forcefun: function of the force, e.g. coriolis

args: other arguments needed for forcefun

Mire.projectforce! — Function

projectforce!(A::AbstractArray{T, 2}, cmat::Array{T, 3}, vs::Array{Array{P, 1}, 1}, N::Integer, forcefun::Function, a::T, b::T, c::T, args...) where {T, P <: Polynomial{T}}

DOCSTRING

#Arguments:

A: pre-allocated array

cmat: pre-cached monomial integration values

vs: basis vectors

N: maximum monomial degree

forcefun: function of the force, e.g. coriolis

args: other arguments needed for forcefun

Missing docstring.

Missing docstring for vel. Check Documenter's build log for details.

Mire.eigenvel — Function

eigenvel(v,α)

Reconstructs velocity u=∑αᵢvᵢ

Missing docstring.

Missing docstring for int_ellipsoid_surface. Check Documenter's build log for details.

Mire.int_monomial_ellipsoid — Function

int_monomial_ellipsoid(i::BigInt, j::BigInt, k::BigInt, a::T, b::T, c::T; dtype::DataType=BigFloat) where T

Integrate a monomial over the ellipsoids volume $\int x^iy^jz^k dV$.

Missing factor

The factor $\pi$ is removed in the integration due to numerical reasons. Remember to reintroduce it when the actual value of the integration is needed!

#Arguments:

i: Exponent of x

j: Exponent of y

k: Exponent of z

a: Semi-axis in x

b: Semi-axis in y

c: Semi-axis in z

Mire.int_polynomial_ellipsoid — Function

int_polynomial_ellipsoid(p, cmat)

DOCSTRING

Mire.inner_product — Function

inner_product(u,v,a,b,c)

Defines inner product in an ellipsoidal volume $\int\langle u,v\rangle dV$.

inner_product(cmat, u, v; thresh=eps())

DOCSTRING

#Arguments:

cmat: DESCRIPTION

u: DESCRIPTION

v: DESCRIPTION

thresh: DESCRIPTION

Missing docstring.

Missing docstring for qg_vel. Check Documenter's build log for details.