Functions

Functions

Setting up the eigen problem

TODO: Description how to set up the problem here.

Mire.assemblehd — Function.

assemblehd(N::Int, a::T, b::T, c::T, Ω ; dtype::DataType=BigFloat, kwargs...) where T

Assemble the sparse matrices of the MHD mode problem. Returns right hand side `A`,
left hand side `B` and basis vectors `vs`.

#Arguments:
- `N`: maximum monomial degree
- `a`: semi-axis x
- `b`: semi-axis y
- `c`: semi-axis z
- `Ω`: rotation vector
- `dtype`: datatype, default `BigFloat` for integration of monomials
- `kwargs`: other keyword arguments passed to lower functions

source

Mire.assemblemhd — Function.

assemblemhd(N::Int, a::T, b::T, c::T, Ω, b0; dtype::DataType=BigFloat, kwargs...) where T

Assemble the sparse matrices of the MHD mode problem. Returns right hand side A, left hand side B and basis vectors vs.

#Arguments:

N: maximum monomial degree

a: semi-axis x

b: semi-axis y

c: semi-axis z

Ω: rotation vector

b0: mean magnetic field vector

dtype: datatype, default BigFloat for integration of monomials

kwargs: other keyword arguments passed to lower functions

source

Low level functions

Functions for low level control of the problem.

Mire.projectforce — Function.

projectforce(N::Integer, vs, forcefun::Function, a::T, b::T, c::T, args...) where T

Allocates new matrix A and fills elements by calling projectforce!(A,vs,N,forcefun,a,b,c, args...).

Cached version: projectforce(N,cmat,vs,forcefun,a,b,c, args...)

where cmat[i,j,k] contains the integrals of monomials xⁱyʲzᵏ.

#Arguments:

N: maximum monomial degree

vs: basis vectors

forcefun: function of the force, e.g. coriolis

a: semi-axis x

b: semi-axis y

c: semi-axis z

args: other arguments needed for forcefun

source

Mire.projectforce! — Function.

projectforce!(A::AbstractArray{T, 2}, cmat::Array{T, 3}, vs::Array{Array{P, 1}, 1}, N::Integer, forcefun::Function, a::T, b::T, c::T, args...; kwargs...) where {T, P <: Polynomial{T}}

DOCSTRING

#Arguments:

A: pre-allocated array

cmat: pre-cached monomial integration values

vs: basis vectors

N: maximum monomial degree

forcefun: function of the force, e.g. coriolis

a: semi-axis x

b: semi-axis y

c: semi-axis z

args: other arguments needed for forcefun

kwargs: other keyword arguments

source

Mire.vel — Function.

vel(N,a,b,c)

Compute all velocity basis vectors for a given maximal degree N (Lebovitz 1989, eq. 41-42)

source

Mire.eigenvel — Function.

eigenvel(N,vs,αs,a,b,c; norm=true)

Reconstructs velocity u, following Vidal & Cebron 2017 eq. (3.5).

If norm keyword is set true the velocity is normalised to satisfy $\int u\cdot u dV=1$.

source

Mire.int_ellipsoid_surface — Function.

Integral over the surface of an ellipsoid.

source

Mire.int_monomial_ellipsoid — Function.

int_monomial_ellipsoid(i::BigInt, j::BigInt, k::BigInt, a::T, b::T, c::T; dtype::DataType=BigFloat) where T

DOCSTRING

#Arguments:

i: DESCRIPTION

j: DESCRIPTION

k: DESCRIPTION

a: DESCRIPTION

b: DESCRIPTION

c: DESCRIPTION

dtype: DESCRIPTION

source

Mire.int_polynomial_ellipsoid — Function.

int_polynomial_ellipsoid(p, cmat)

DOCSTRING

source

Mire.inner_product — Function.

inner_product(u,v,a,b,c)

Defines inner product in an ellipsoidal volume $\int\langle u,v\rangle dV$.

source

inner_product(cmat, u, v; thresh=eps())

DOCSTRING

#Arguments:

cmat: DESCRIPTION

u: DESCRIPTION

v: DESCRIPTION

thresh: DESCRIPTION

source