Functions

Functions

Setting up the eigen problem

TODO: Description how to set up the problem here.

Mire.assemblehd — Function.

assemblehd(N::Int, a::T, b::T, c::T, Ω ; dtype::DataType=BigFloat, kwargs...) where T

Assemble the sparse matrices of the MHD mode problem. Returns right hand side A, left hand side B and basis vectors vs.

#Arguments:

N: maximum monomial degree

a: semi-axis x

b: semi-axis y

c: semi-axis z

Ω: rotation vector

dtype: datatype, default BigFloat for integration of monomials

kwargs: other keyword arguments passed to lower functions

source

Mire.assemblemhd — Function.

assemblemhd(N::Int, a::T, b::T, c::T, Ω, b0; dtype::DataType=BigFloat, kwargs...) where T

Assemble the sparse matrices of the MHD mode problem. Returns right hand side A, left hand side B and basis vectors vs.

#Arguments:

N: maximum monomial degree

a: semi-axis x

b: semi-axis y

c: semi-axis z

Ω: rotation vector

b0: mean magnetic field vector

dtype: datatype, default BigFloat for integration of monomials

kwargs: other keyword arguments passed to lower functions

source

Mire.assemblemhd_hybrid — Function.

assemblehd_hybrid(N2D::Int, N3D::Int, a::T, b::T, c::T, Ω ; dtype::DataType=BigFloat, kwargs...) where T

Assemble the sparse matrices of the hybrid QG and 3D MHD mode problem. Returns right hand side A,left hand side B and basis vectors vs and vs_qg.

#Arguments:

N2D: maximum monomial degree of QG velocity

N3D: maximum monomial degree of 3D magnetic field

a: semi-axis x

b: semi-axis y

c: semi-axis z

Ω: rotation vector

dtype: datatype, default BigFloat for integration of monomials

kwargs: other keyword arguments passed to lower functions

source

Low level functions

Functions for low level control of the problem.

Mire.projectforce — Function.

projectforce(N::Integer,cmat::Array{T,3},vs::Array{Array{P,1},1},
forcefun::Function,a::T,b::T,c::T, args...; kwargs...) where {T, P <: Polynomial{T}}

Allocates new matrix A and fills elements by calling projectforce!(A,cmat,vs,forcefun, args...; kwargs...)

where cmat[i,j,k] contains the integrals of monomials xⁱyʲzᵏ.

#Arguments:

N: maximum monomial degree

vs: basis vectors

forcefun: function of the force, e.g. coriolis

args: other arguments needed for forcefun

source

Mire.projectforce! — Function.

projectforce!(A::AbstractArray{T, 2}, cmat::Array{T, 3}, vs::Array{Array{P, 1}, 1}, N::Integer, forcefun::Function, a::T, b::T, c::T, args...; kwargs...) where {T, P <: Polynomial{T}}

DOCSTRING

#Arguments:

A: pre-allocated array

cmat: pre-cached monomial integration values

vs: basis vectors

N: maximum monomial degree

forcefun: function of the force, e.g. coriolis

args: other arguments needed for forcefun

kwargs: other keyword arguments

source

Mire.vel — Function.

vel(N,a,b,c)

Compute all velocity basis vectors for a given maximal degree N (Lebovitz 1989, eq. 41-42)

source

Mire.eigenvel — Function.

eigenvel(v,α)

Reconstructs velocity u=∑αᵢvᵢ

source

Missing docstring.

Missing docstring for int_ellipsoid_surface. Check Documenter's build log for details.

Mire.int_monomial_ellipsoid — Function.

int_monomial_ellipsoid(i::BigInt, j::BigInt, k::BigInt, a::T, b::T, c::T; dtype::DataType=BigFloat) where T

DOCSTRING

#Arguments:

i: DESCRIPTION

j: DESCRIPTION

k: DESCRIPTION

a: DESCRIPTION

b: DESCRIPTION

c: DESCRIPTION

dtype: DESCRIPTION

source

Mire.int_polynomial_ellipsoid — Function.

int_polynomial_ellipsoid(p, cmat)

DOCSTRING

source

Mire.inner_product — Function.

inner_product(u,v,a,b,c)

Defines inner product in an ellipsoidal volume $\int\langle u,v\rangle dV$.

source

inner_product(cmat, u, v; thresh=eps())

DOCSTRING

#Arguments:

cmat: DESCRIPTION

u: DESCRIPTION

v: DESCRIPTION

thresh: DESCRIPTION

source

Mire.qg_vel — Function.

qg_vel(n::Integer,m::Integer,a::T,b::T,c::T) where T

Generate basis vector $\mathbf{u}=\nabla(h^3x^ny^m)\times\nabla(z/h)$

source